update
- move LightningTransformer
- implement suggestions
