+++ src/lightning/pytorch/demos/__init__.py
@@ -1,2 +1,1 @@
-from lightning.pytorch.demos.lightning_transformer import LightningTransformer  # noqa: F401
 from lightning.pytorch.demos.transformer import Transformer, WikiText2  # noqa: F401
