+++ src/lightning/pytorch/demos/transformer.py
@@ -17,6 +17,8 @@ from torch import Tensor
 from torch.nn.modules import MultiheadAttention
 from torch.utils.data import Dataset
 
+import lightning.pytorch as pl
+
 if hasattr(MultiheadAttention, "_reset_parameters") and not hasattr(MultiheadAttention, "reset_parameters"):
     # See https://github.com/pytorch/pytorch/issues/107909
     MultiheadAttention.reset_parameters = MultiheadAttention._reset_parameters
@@ -161,3 +163,22 @@ def tokenize(path: Path) -> Tuple[Tensor, Dictionary]:
             idss.append(torch.tensor(ids).type(torch.int64))
 
     return torch.cat(idss), dictionary
+
+
+class LightningTransformer(pl.LightningModule):
+    def __init__(self, vocab_size: int) -> None:
+        super().__init__()
+        self.model = Transformer(vocab_size=vocab_size)
+
+    def forward(self, batch: Tuple[Tensor, Tensor]) -> Tensor:
+        input, target = batch
+        return self.model(input, target)
+
+    def training_step(self, batch: Tuple[Tensor, Tensor], batch_idx: int) -> Tensor:
+        input, target = batch
+        output = self(batch)
+        loss = torch.nn.functional.nll_loss(output, target.view(-1))
+        return loss
+
+    def configure_optimizers(self) -> torch.optim.Optimizer:
+        return torch.optim.SGD(self.model.parameters(), lr=0.1)
